+++ modules/utils/src/index.ts
@@ -1,4 +1,3 @@
-export * from "./attestations";
 export * from "./bigNumbers";
 export * from "./channelSigner";
 export * from "./comparisons";
@@ -9,6 +8,7 @@ export * from "./delay";
 export * from "./eip712";
 export * from "./env";
 export * from "./fs";
+export * from "./graphTransfer";
 export * from "./hexStrings";
 export * from "./identifiers";
 export * from "./json";
